requirements.txt
Flask==3.0.3
pymongo==4.6.3
PyJWT==2.8.0
bcrypt==4.1.2
python-dotenv==1.0.1
playwright==1.42.0
requests==2.31.0
Pillow==10.4.0
Flask-Limiter==3.8.0
markdown==3.6
